% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC.R
\name{ICC}
\alias{ICC}
\title{Intraclass Correlation Coefficient (ICC) for a set of athlete measurements}
\usage{
ICC(subject, trial, ...)
}
\arguments{
\item{subject}{The vector of athletes who recorded the results for each metric}

\item{trial}{The vector that represents which trial each measurement came from. In order for the function to work, each athlete
must have completed the same number of trials for each metric.}

\item{...}{Numeric vectors that represent the metrics for which the ICC output should be given. These vectors hold the scores that
each athlete recorded for each respective metric (at least one metric must be passed to the function).}
}
\value{
A list, with its contents being the ICC output (generated by the ICC from the psych package, with the statistics for the
  six types of ICC's), is returned
}
\description{
Computes the ICC for each vector of measurements that is passed to the function, for the vectors that are the first
  two arguments called to the function (the subject vector and the trial vector, in that order).
}
\references{
Shrout, P. E., &amp; Fleiss, J. L. (1979). Intraclass correlations: Uses in assessing rater reliability. Psychological
  Bulletin, 86(2), 420-428.
}
