% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOA.R
\name{LoA}
\alias{LoA}
\title{The Bland-Altman Limits of Agreement (LoA) for a set of athlete measurements}
\source{
Hopkins, W. G. (2000). Measures of Reliability in Sports Medicine and Science. Sports Medicine, 30(5), 375-381.
}
\usage{
LoA(subject, trial, ..., confidence = 0.95)
}
\arguments{
\item{subject}{The vector of athletes who recorded the results for each metric (can be a numeric or factor variable)}

\item{trial}{The vector that represents which trial each measurement came from}

\item{...}{Numeric vectors that represent the metrics for which the CV should be computed. These vectors hold the scores that
each athlete recorded for each respective metric (at least one metric must be passed to the function)}

\item{confidence}{The degree of confidence the user wants to have that an improvement exceeding the MDD can be interpreted as
real change, and not the result of measurement error. Set to a default value of 0.95, this parameter is used to calculate the
corresponding critical value from the standard normal distribution to which we compare the RCI.}
}
\value{
A table, with each metric being its own column and its entry being the LoA of that metric, is the output of this function
}
\description{
Computes the LoA for each vector of measurements that is passed to the function, for the vector of subjects in the
  first argument
}
