% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDIVIDUAL_CHANGE.R
\name{MDC}
\alias{MDC}
\title{Minimal Detectable Change (MDC) for a set of athlete measurements}
\usage{
MDC(
  subject,
  ...,
  reliability,
  confidence = 0.95,
  method = c("AVG", "MAX", "MIN")
)
}
\arguments{
\item{subject}{The vector of athletes who recorded the results for each metric (can be a numeric or factor variable)}

\item{...}{Numeric vectors that represent the metrics for which the SEM should be computed. These vectors hold the scores that
each athlete recorded for each respective metric (at least one metric must be passed to the function).}

\item{reliability}{A vector of the measures of reliability (i.e. the ICC's) for each of the metrics included in the "..."
argument. It is suggested that the reliability measure chosen be the ICC, but it can be any reliability measure, as long as the
measure chosen lies on a 0-1 scale (with 1 indicating higher reliability). This vector must contain the same number of
elements as the number of metrics that have been passed to the function in the "..." argument, and the reliability values must
appear in the same order as the metrics appear in the "..." argument.}

\item{confidence}{The degree of confidence the user wants to have that an improvement exceeding the MDC can be interpreted as
real change, and not the result of measurement error. Set to a default value of 0.95, this parameter is used to calculate the
corresponding critical value from the standard normal distribution to which we compare the RCI.}

\item{method}{The user's choice of how the between-athlete SD should be computed in the formula for the MDC. If set to AVG, each
athlete's values will be averaged before the SD of these between-athlete averages is computed. If MAX is selected, then only
the highest value each athlete records will be included in the computation of the between-athlete SD. Conversely, if the user
indicates MIN, then only the lowest value each athlete records will be used to compute the between-athlete SD.}
}
\value{
A table, with each metric being its own column and its entry being the MDC of that metric, is the output of this function
}
\description{
Computes the MDC for each metric that is passed to the function as a vector of athlete measurements
}
\references{
Riemann, B. L., &amp; Lininger, M. R. (2018). Statistical Primer for Athletic Trainers: The Essentials of
  Understanding Measures of Reliability and Minimal Important Change. Journal of Athletic Training, 53(1), 98-103.
}
