% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERFORMANCESCIENCERELIABILITY.R
\docType{package}
\name{PerfSciReliability}
\alias{PerfSciReliability}
\title{PerfSciReliability}
\description{
A package for computing various measures pertaining to reliability of performance science metrics. It contains
  functions for the Typical Error (TE), Coefficient of Variation (CV), Smallest Worthwhile Change (SWC), Minimal Detectable Change
  (MDC), Limits of Agreement (LoA), Standard Error of Measurement (SEM), Intraclass Correlation Coefficient (ICC), and Reliable
  Change Index (RCI).
}
