% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RELIABLE_CHANGE.R
\name{prob_rel_chng}
\alias{prob_rel_chng}
\title{Probability of a given change for a given metric being reliable from trial to trial, for each athlete in the data}
\usage{
prob_rel_chng(subject, initial, final, ..., reliability)
}
\arguments{
\item{subject}{The subjects for which the probabilities of reliable change should be computed}

\item{initial}{The set of measurements to be considered the "before" measurements}

\item{final}{The set of measurements to be considered the "after" measurements}

\item{...}{Numeric vectors that represent the metrics for which the probability of the given changes being reliable should be
computed. These vectors hold the scores that each athlete recorded for each respective metric (at least one metric must be
passed to the function).}

\item{reliability}{A vector of the measures of reliability (i.e. the ICC's) for each of the metrics included in the
"..." argument. This vector must contain the same number of elements as the number of metrics that have been
passed to the function in the "..." argument, and the reliability values must appear in the same order as the
metrics appear in the "..." argument.}
}
\value{
A table for each metric, with each row being an athlete and each entry being the probability that the change that has
  occurred is reliable, based on the RCI calculation.
}
\description{
For each change that an athlete records from one trial to another, the probability that the change is reliable is
  generated, along with asterisks for athletes that have achieved given probabilities of change.
}
\references{
Ferger, K., & Büsch, D. (2018). Individual measurement of performance change in sports. Deutsche Zeitschrift Für
  Sportmedizin, 2018(02), 45-52.
}
