% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RELIABLE_CHANGE.R
\name{RCI}
\alias{RCI}
\title{Reliable Change Index (RCI) for a set of athlete measurements}
\usage{
RCI(subject, baseline, ..., reliability, confidence = 0.95)
}
\arguments{
\item{subject}{The subjects on which the RCI test should be run}

\item{baseline}{The measurements to be considered the "baseline" results for the calculation of the RCI}

\item{...}{Numeric vectors that represent the metrics for which the RCI should be computed. These vectors hold the
scores that each athlete recorded for each respective metric (at least one metric must be passed to the function).}

\item{reliability}{A vector of the measures of reliability (i.e. the ICC's) for each of the metrics included in the "..."
argument. It is suggested that the reliability measure chosen be the ICC, but it can be any reliability measure, as long as the
measure chosen lies on a 0-1 scale (with 1 indicating higher reliability). This vector must contain the same number of
elements as the number of metrics that have been passed to the function in the "..." argument, and the reliability values must
appear in the same order as the metrics appear in the "..." argument.}

\item{confidence}{The degree of confidence the user wants to have that the lower and upper bounds given can be
interpreted as real change, and not the result of random or measurement error. Set to a default value of 0.95, this
parameter is used to calculate the corresponding critical value from the standard normal distribution that is used to compute
the lower and upper bounds of the RCI.}
}
\value{
A table for each metric, with each row being an athlete and two columns, one for the RCI lower bound and the other for
  the RCI upper bound.
}
\description{
For each metric passed to the function, vectors indicating the lower and upper bounds of the RCI of that metric are
  generated, with the length of each vector equaling the number of athletes passed to the function
}
\references{
Ferger, K., & Büsch, D. (2018). Individual measurement of performance change in sports. Deutsche Zeitschrift Für
  Sportmedizin, 2018(02), 45-52.
}
