% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSOLUTE_RELIABILITY.R
\name{TE}
\alias{TE}
\title{Typical Error (TE) for a set of athlete measurements}
\usage{
TE(subject, trial, ...)
}
\arguments{
\item{subject}{The vector of athletes who recorded the results for each metric (can be a numeric or factor variable)}

\item{trial}{The vector that represents which trial each measurement came from}

\item{...}{Numeric vectors that represent the metrics for which the CV should be computed. These vectors hold the scores that
each athlete recorded for each respective metric (at least one metric must be passed to the function)}
}
\value{
A table, with each metric being its own column and its entry being the TE of that metric, is the output of this function
}
\description{
Computes the TE for each vector of measurements that is passed to the function, for the vector of subjects in the
  first argument
}
\references{
Hopkins, W. G. (2000). Measures of Reliability in Sports Medicine and Science. Sports Medicine, 30(5), 375-381.
}
