% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSOLUTE_RELIABILITY.R
\name{CV}
\alias{CV}
\title{Coefficient of Variation (CV) for a set of athlete measurements, expressed as a percentage of the mean}
\usage{
CV(subject, trial, ...)
}
\arguments{
\item{subject}{The vector of athletes who recorded the results for each metric (can be a numeric or factor variable)}

\item{...}{Numeric vectors that represent the metrics for which the CV should be computed. These vectors hold the scores that
each athlete recorded for each respective metric (at least one metric must be passed to the function).}
}
\value{
A table, with each metric being its own column and its entry being the CV of that metric, is the output of this function
}
\description{
Computes the CV for each vector of measurements that is passed to the function, for the vector of subjects in the
  first argument
}
\references{
Hopkins, W. G. (2000). Measures of Reliability in Sports Medicine and Science. Sports Medicine, 30(5), 375-381.
}
