% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC.R
\name{ICC_long}
\alias{ICC_long}
\title{Intraclass Correlation Coefficient (ICC) for a set of athlete measurements, with the data in long format}
\usage{
ICC_long(subject, trial, ...)
}
\arguments{
\item{subject}{The vector of athletes who recorded the results for each metric (can be a numeric or factor variable)}

\item{trial}{The vector that represents which trial each measurement came from}

\item{...}{Numeric vectors that represent the metrics for which the ICC output should be given. These vectors hold the scores that
each athlete recorded for each respective metric (at least one metric must be passed to the function).}
}
\value{
A list, with its contents being the ICC output (containing the six types of ICC's, as generated by psych::ICC()), is
  returned
}
\description{
Computes the ICC for each vector of measurements that is passed to the function, for the subject and trial vectors
}
\references{
Shrout, P. E., &amp; Fleiss, J. L. (1979). Intraclass correlations: Uses in assessing rater reliability. Psychological
  Bulletin, 86(2), 420-428.
}
