% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWC,MDC.R
\name{MDC}
\alias{MDC}
\title{Minimal Detectable Change (MDC) for a set of athlete measurements}
\usage{
MDC(subject, trial, ..., ICC, confidence = 0.95)
}
\arguments{
\item{subject}{The vector of athletes who recorded the results for each metric (can be a numeric or factor variable)}

\item{trial}{The vector that represents which trial each measurement came from}

\item{...}{Numeric vectors that represent the metrics for which the SEM should be computed. These vectors hold the scores that
each athlete recorded for each respective metric (at least one metric must be passed to the function).}

\item{ICC}{A vector of the ICC's for each of the metrics included in the "..." argument. This vector must contain the same number
of elements as the number of metrics that have been passed to the function in the "..." argument, and the reliability values
must appear in the same order as the metrics appear in the "..." argument.}

\item{confidence}{The degree of confidence the user wants to have that an improvement exceeding the MDC can be interpreted as
real change, and not the result of measurement error, based on the standard normal distribution. The default value is 0.95.}
}
\value{
A data frame, with the name of each metric situated above its calculated MDC
}
\description{
Computes the MDC for each metric that is passed to the function as a vector of athlete measurements, for the
  subject and trial vectors, the ICC's of each metric, and the desired confidence level
}
\references{
Riemann, B. L., &amp; Lininger, M. R. (2018). Statistical Primer for Athletic Trainers: The Essentials of
  Understanding Measures of Reliability and Minimal Important Change. Journal of Athletic Training, 53(1), 98-103.
}
