% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROBRC.R
\name{PROBRC}
\alias{PROBRC}
\title{Probability of a given change for a given metric being reliable from trial to trial, for each athlete in the data}
\usage{
PROBRC(subject, trial, ..., ICC, initial, final)
}
\arguments{
\item{subject}{The vector of athletes who recorded the results for each metric (can be a numeric or factor variable)}

\item{trial}{The vector that represents which trial each measurement came from}

\item{...}{Numeric vectors that represent the metrics for which the probability of the given changes being reliable should be
computed. These vectors hold the scores that each athlete recorded for each respective metric (at least one metric must be
passed to the function).}

\item{ICC}{A vector of the ICC's for each of the metrics included in the "..." argument. This vector must contain the same number
of elements as the number of metrics that have been passed to the function in the "..." argument, and the reliability values
must appear in the same order as the metrics appear in the "..." argument.}

\item{initial}{The trial to be considered the "before" (or baseline) measurements}

\item{final}{The trial to be considered the "after" measurements (the trial we are comparing to the baseline)}
}
\value{
A data frame, with the subjects as rows and the metrics as columns, and each entry representing the probability that the
  final measurement was a reliable change from the individual measurement.
}
\description{
For each change that an athlete records from one trial to another, the probability that the change is reliable is
  generated, along with asterisks for athletes that have achieved given probabilities of change.
}
\references{
Ferger, K., & Büsch, D. (2018). Individual measurement of performance change in sports. Deutsche Zeitschrift Für
  Sportmedizin, 2018(02), 45-52.
}
