% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psr.R
\docType{package}
\name{psr}
\alias{psr}
\title{psr}
\description{
A package for computing various measures relating to reliability of performance science metrics. It
  contains functions to compute the Typical Error (TE), Coefficient of Variation (CV), Standard Error of Measurement
  (SEM), Smallest Worthwhile Change (SWC), Minimal Detectable Change (MDC), Intraclass Correlation Coefficient (ICC),
  Standard Ten Scores (STEN), and Probability of a Reliable Change (PROBRC).
}
