% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSOLUTE_RELIABILITY.R
\name{SEM}
\alias{SEM}
\title{Standard Error of Measurement (SEM) for a set of athlete measurements}
\usage{
SEM(subject, trial, ..., reliability, method = c("AVG", "MAX", "MIN"))
}
\arguments{
\item{subject}{The vector of athletes who recorded the results for each metric (can be a numeric or factor variable)}

\item{...}{Numeric vectors that represent the metrics for which the SEM should be computed. These vectors hold the scores that
each athlete recorded for each respective metric (at least one metric must be passed to the function).}

\item{method}{The user's choice of how the between-athlete SD should be computed in the formula for the SEM. If set to AVG, each
athlete's values will be averaged before the SD of these between-athlete averages is computed. If MAX is selected, then only
the highest value each athlete records will be included in the computation of the between-athlete SD. Conversely, if the user
indicates MIN, then only the lowest value each athlete records will be used to compute the between-athlete SD.}

\item{ICC}{A vector of the ICC's for each of the metrics included in the "..." argument. This vector must contain the same number of
elements as the number of metrics that have been passed to the function in the "..." argument, and the reliability values must
appear in the same order as the metrics appear in the "..." argument.}
}
\value{
A table, with each metric being its own column and its entry being the SEM of that metric, is the output of this function
}
\description{
Computes the SEM for each metric that is passed to the function as a vector of measurements
}
\references{
Atkinson, G., & Nevill, A. M. (1998). Statistical Methods For Assessing Measurement Error (Reliability) in Variables
  Relevant to Sports Medicine. Sports Medicine, 26(4), 217-238.
}
