% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STANDARDIZED_SCORES.R
\name{STEN}
\alias{STEN}
\title{Standard Ten (STEN) Scores for a set of athlete measurements}
\usage{
STEN(subject, ...)
}
\arguments{
\item{subject}{The vector of athletes who recorded the results for each metric (can be a numeric or factor variable)}

\item{...}{Numeric vectors that represent the metrics which scores should be computed to STEN scores. These vectors hold the
scores that each athlete recorded for each respective metric (at least one metric must be passed to the function).}
}
\value{
The measurements of the metrics are converted from their respective original units of measurement to the unitless STEN
  score that the measurement represents.
}
\description{
Converts each score of each metric passed to the function into its corresponding STEN score, which is a way of
  putting scores that have different units on the same scale, namely that which has a minimum value of 1 and a maximum value of 10
  for all metrics. This serves the main purpose of allowing for easy comparison of an athlete's scores for different metrics that
  might be measured in different units (so that it can be identified which metrics are areas of strength and which are areas of
  weakness for the athlete). The STEN method fulfills the same purpose as standardizing the scores (i.e. converting all of the
  scores to z-scores), but has the advantage of potentially being easier to understand for some practitioners and coaches.
}
