% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDIVIDUAL_CHANGE.R
\name{SWC}
\alias{SWC}
\title{Smallest Worthwhile Change (SWC) for a set of athlete measurements}
\usage{
SWC(subject, ..., method = c("AVG", "MAX", "MIN"))
}
\arguments{
\item{subject}{The vector of athletes who recorded the results for each metric (can be a numeric or factor variable)}

\item{...}{Numeric vectors that represent the metrics for which the SWC should be computed. These vectors hold the scores that
each athlete recorded for each respective metric (at least one metric must be passed to the function)}

\item{method}{The user's choice of how the between-athlete SD should be computed in the formula for the SWC. If set to AVG, each
athlete's values will be averaged before the SD of these between-athlete averages is computed. If MAX is selected, then only
the highest value each athlete records will be included in the computation of the between-athlete SD. Conversely, if the user
indicates MIN, then only the lowest value each athlete records will be used to compute the between-athlete SD.}
}
\value{
A table, with each metric being its own column and its entry being the SWC of that metric, is the output of this function
}
\description{
Computes the SWC for each vector of measurements corresponding to a vector of athletes that are its arguments
}
\references{
Bernards, J., Sato, K., Haff, G., & Bazyler, C. (2017). Current Research and Statistical Practices in Sport
  Science and a Need for Change. Sports, 5(4), 87.
}
